Struct bevy_datasize::prelude::MemoryStats

pub struct MemoryStats {
    pub count: usize,
    pub total_stack_bytes: usize,
    pub total_heap_bytes: usize,
}

Memory usage statistics for a single data type.

Example

#[derive(DataSize)]
struct SomeData {
    data: Vec<u8>,
}

assert_eq!(std::mem::size_of::<SomeData>(), 24);

let values = vec![
    SomeData { data: vec![0; 100] },
    SomeData { data: vec![0; 100] },
];

let stats = MemoryStats::from_values(values.iter());

assert_eq!(stats.count, 2);
assert_eq!(stats.total_stack_bytes, 48);
assert_eq!(stats.total_heap_bytes, 200);
assert_eq!(stats.total_bytes(), 248);
assert_eq!(format!("{stats}"), "2 (248 B)")

Fields
count: usize

The total number of instances of this data type.

total_stack_bytes: usize

The total number of “stack” bytes used by instances of this data type.

See stack_size_of for details on the meaning of this quantity.

total_heap_bytes: usize

The estimated total number of bytes used by instances of this data type.

See heap_size_of for details on the meaning of this quantity.

Implementations

impl MemoryStats

pub fn total_bytes(&self) -> usize

Returns the sum of total_stack_bytes and total_heap_bytes for self.

pub fn from_value<T>(value: &T) -> Self where
T: Any + DataSize,

Returns the computed memory statistics for a single value.

pub fn from_value_with_estimator<T, E>(value: &T, estimator: &E) -> Self where
T: Any,
E: DataSizeEstimator<T>,

Returns the computed memory statistics for a single value using a specific DataSizeEstimator.

pub fn from_values<'a, T, I>(values: I) -> Self where
T: Any + DataSize,
I: IntoIterator<Item = &'a T>,

Returns the computed memory statistics for a collection of values.

pub fn from_values_with_estimator<'a, T, E, I>(values: I, estimator: &E) -> Self where
T: Any,
E: DataSizeEstimator<T>,
I: IntoIterator<Item = &'a T>,

Returns the computed memory statistics for a collection of values.

pub fn stack_size_of<T>(value: &T) -> usize where
T: Any,

Returns the “stack” size of the given value.

This quantity represents how many bytes the type would take up on the stack if were allocated there.

This quantity is exact and is computed using std::mem::size_of.

pub fn heap_size_of<T>(value: &T) -> usize where
T: DataSize,

Returns the estimated heap size of the given value.

This quantity represents how many bytes the type occupies apart from the immediate bytes of its fields.

This quantity is estimated and may not be 100% accurate for all types.

pub fn heap_size_of_with_estimator<T, E>(value: &T, estimator: &E) -> usize where
E: DataSizeEstimator<T>,

Returns the estimated heap size of the given value using a specific DataSizeEstimator.

pub fn total_size_of<T>(value: &T) -> usize where
T: Any + DataSize,

Returns the estimated total size of the given value.

This quantity is the sum of stack_size_of and heap_size_of.

pub fn total_size_of_with_estimator<T, E>(value: &T, estimator: &E) -> usize where
T: Any,
E: DataSizeEstimator<T>,

Returns the estimated total size of the given value using a specific DataSizeEstimator.